+++ test/zmqComms/test_zmqSerialize.py
@@ -124,7 +124,10 @@ class ZmqSerializeTest(unittest.TestCase):
       "foo": {
         "descriptor": "foodesc", 
         "type": "int", 
-        "value": "arg_required"
+        "value": null, 
+        "tags": [
+          "required"
+        ]
       }, 
       "bar": {
         "descriptor": "bardesc", 
@@ -327,7 +330,10 @@ class ZmqSerializeTest(unittest.TestCase):
           "PC_BIT_CAP": {
             "descriptor": "Which encoders to capture", 
             "type": "int", 
-            "value": "arg_required"
+            "value": null, 
+            "tags": [
+              "required"
+            ]
           }, 
           "PC_TSPRE": {
             "descriptor": "What time units for capture",